Update to rc.2; added filing-chart component placeholder with animation
