+++ .eslintrc
@@ -1,4 +1,5 @@
 {
+    "ecmaVersion": 6,
     "rules": {
         "semi": ["error", "always"],
         "quotes": ["error", "double"]
