+++ package.json
@@ -67,15 +67,14 @@
     "express:prod": "NODE_ENV=production node server/main.js"
   },
   "dependencies": {
-    "@angular/common":  "2.0.0-rc.1",
-    "@angular/compiler":  "2.0.0-rc.1",
-    "@angular/core":  "2.0.0-rc.1",
-    "@angular/http":  "2.0.0-rc.1",
-    "@angular/platform-browser":  "2.0.0-rc.1",
-    "@angular/platform-browser-dynamic":  "2.0.0-rc.1",
-    "@angular/router":  "3.0.0-alpha.3",
-    "@angular/router-deprecated":  "2.0.0-rc.1",
-    "@angular/upgrade":  "2.0.0-rc.1",
+    "@angular/common":  "2.0.0-rc.2",
+    "@angular/compiler":  "2.0.0-rc.2",
+    "@angular/core":  "2.0.0-rc.2",
+    "@angular/http":  "2.0.0-rc.2",
+    "@angular/platform-browser":  "2.0.0-rc.2",
+    "@angular/platform-browser-dynamic":  "2.0.0-rc.2",
+    "@angular/router":  "3.0.0-alpha.6",
+    "@angular/upgrade":  "2.0.0-rc.2",
     
     "core-js": "^2.4.0",
     "normalize.css": "^4.1.1",
