+++ src/app/compare/compare.routes.ts
@@ -1,5 +1,5 @@
 import { CompareComponent }   from './compare.component';
 
 export const CompareRoutes = [
-  { path: '/compare',  component: CompareComponent }
+  { path: 'compare',  component: CompareComponent }
 ];
