+++ src/app/launch/launch.routes.ts
@@ -1,5 +1,5 @@
 import { LaunchComponent }   from './launch.component';
 
 export const LaunchRoutes = [
-  { path: '/launch',  component: LaunchComponent, index: true }
+  { path: 'launch',  component: LaunchComponent, index: true }
 ];
