+++ src/app/model/symbol.ts
@@ -1,0 +1,8 @@
+export interface Symbol {
+    Symbol: string;
+    Name: string;
+    Sector: string;
+    Industry: string;
+    exchange: string;
+    SymbolHREF: string;
+}
